final backend requests and exceptions

+++ Eshop/src/main/java/Eshop/demo/product/ProductService.java
@@ -4,8 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.aggregation.Aggregation;
 import org.springframework.data.mongodb.core.aggregation.AggregationResults;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,7 +24,9 @@ public class ProductService {
     }
 
     public void saveProduct(Product product){
-         productRepository.save(product);
+    if(product.getId() == null || product.getCategory() == null || product.getName() == null || product.getPrice() == null)
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid arguments");
+    productRepository.save(product);
     }
 
     public void cleanup(){
@@ -36,9 +40,11 @@ public class ProductService {
         return productDTOList;
     }
     public Product getProductDetailsById(String id){
+        if(productRepository.findById(id).isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found");
         return productRepository.findById(id).orElse(null);
     }
     public void deleteProduct(String id){
+        if(productRepository.findById(id).isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found");
         productRepository.deleteById(id);
     }
